Set musicId during initialization
We should set the `musicId` during the initialization of the
component, instead of only changing it when fetching the music
and lyrics. This is because

1. musicId will never change through out the life of the
MusicPlayer (unless you support "Next Song" button), so we should
set it *as early as possible*, i.e. during the initialization
phase.

2. setState is *asynchronous*. You used to put these 3 async actions
together in `componentDidMount`:
  - setState
  - fetch music
  - fetch lyric

  The order of which one got executed first completely random. So
  even if you setState first, it's possible that when fetching
  music is called later, it can't get the correct `musicId`